* PACKET_ID 와 PacketId 타입을 왔다 갔다 하기 싫어서 PACKET_ID로 통일 * 액터에 패킷 핸들러를 등록할 수 있도록 함수 수정 * 패킷 핸들러와 패킷 핸들러 등록 전용 파일 추가   * 패킷 핸들러 예시로 HandlePing() 추가

+++ ActorModelServer/ActorModelServer/Session.cpp
@@ -64,7 +64,7 @@ bool Session::SendPacket(IPacket& packet)
 		return false;
 	}
 
-	*buffer << packet.GetPacketId();
+	*buffer << static_cast<unsigned int>(packet.GetPacketId());
 	buffer->WriteBuffer((char*)(&packet) + 8, packet.GetPacketSize());
 
 	return DoSend();
